feat: 将header切换逻辑统一转移到app.vue
